fake_news_backend/requirements.txt
python-dotenv
openai
beautifulsoup4
requests
pydantic
fastapi
uvicorn
httpx